# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "wavynote",
            "email": "chelseafandev@gmail.com"
        }
    },
    "paths": {
...
        "/profile/duplicate": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "회원 가입 시 ID 중복 체크",
                "tags": [
                    "SignUp/SignIn 페이지"
                ],
                "summary": "회원 가입 시 ID 중복 체크",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restapi.DefaultResponse"
                        }
                    },
                    "400": {
                        "description": "요청에 포함된 파라미터 값이 잘못된 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response400"
                        }
                    },
                    "401": {
                        "description": "인증에 실패한 경우이며, 실패 사유가 전달됩니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response401"
                        }
                    },
                    "404": {
                        "description": "요청한 리소스가 서버에 존재하지 않는 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response404"
                        }
                    },
                    "500": {
                        "description": "요청을 처리하는 과정에서 서버에 문제가 발생한 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response500"
                        }
                    }
                }
            }
        },
        "/profile/signin": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "로그인 요청",
                "tags": [
                    "SignUp/SignIn 페이지"
                ],
                "summary": "로그인 요청",
                "parameters": [
                    {
                        "description": "로그인 요청",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/restapi.SignInRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "랜딩페이지 출력 시 필요한 정보",
                        "schema": {
                            "$ref": "#/definitions/restapi.LandingPageResonse"
                        }
                    },
                    "400": {
                        "description": "요청에 포함된 파라미터 값이 잘못된 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response400"
                        }
                    },
                    "401": {
                        "description": "인증에 실패한 경우이며, 실패 사유가 전달됩니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response401"
                        }
                    },
                    "404": {
                        "description": "요청한 리소스가 서버에 존재하지 않는 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response404"
                        }
                    },
                    "500": {
                        "description": "요청을 처리하는 과정에서 서버에 문제가 발생한 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response500"
                        }
                    }
                }
            }
        },
        "/profile/signup": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "회원 가입 요청",
                "tags": [
                    "SignUp/SignIn 페이지"
                ],
                "summary": "회원 가입 요청",
                "parameters": [
                    {
                        "description": "회원 가입 요청 정보",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/restapi.SignUpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "랜딩페이지 출력 시 필요한 정보",
                        "schema": {
                            "$ref": "#/definitions/restapi.LandingPageResonse"
                        }
                    },
                    "400": {
                        "description": "요청에 포함된 파라미터 값이 잘못된 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response400"
                        }
                    },
                    "401": {
                        "description": "인증에 실패한 경우이며, 실패 사유가 전달됩니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response401"
                        }
                    },
                    "404": {
                        "description": "요청한 리소스가 서버에 존재하지 않는 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response404"
                        }
                    },
                    "500": {
                        "description": "요청을 처리하는 과정에서 서버에 문제가 발생한 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response500"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "restapi.DefaultResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "description": "실패 시 반환하는 에러 메시지",
                    "type": "string"
                },
                "result": {
                    "description": "요청에 대한 처리 성공/실패 여부",
                    "type": "string",
                    "example": "true"
                }
            }
        },
...
        "restapi.LandingPageResonse": {
            "type": "object",
            "properties": {
                "folder_id": {
                    "description": "랜딩 페이지에서 보여줄 폴더의 고유 id 값",
                    "type": "string",
                    "example": "283744fe-7074-4ffa-abc6-dae8cb786058"
                },
                "user_id": {
                    "description": "사용자 id",
                    "type": "string",
                    "example": "somebody@naver.com"
                }
            }
        },
...
        "restapi.Response400": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "msg": {
                    "type": "string",
                    "example": "Bad request"
                }
            }
        },
        "restapi.Response401": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 401
                },
                "msg": {
                    "type": "string",
                    "example": "Fail to parse Authorization header"
                }
            }
        },
        "restapi.Response404": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "msg": {
                    "type": "string",
                    "example": "Not exist the target resource in our server"
                }
            }
        },
        "restapi.Response500": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "msg": {
                    "type": "string",
                    "example": "Internal server error"
                }
            }
        },
...
        "restapi.SignInRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "로그인시 입력한 ID",
                    "type": "string",
                    "example": "somebody@naver.com"
                },
                "password": {
                    "description": "로그인시 입력한 패스워드",
                    "type": "string",
                    "example": "04F0BF9AB5963706415A0EC4C0E8167EBC8640B0B9A40870CDC39BE8A8DA813B"
                }
            }
        },
        "restapi.SignUpRequest": {
            "type": "object",
            "properties": {
                "emoji": {
                    "description": "회원 가입시 입력한 이모지",
                    "type": "string",
                    "example": "😎"
                },
                "id": {
                    "description": "회원 가입시 입력한 ID",
                    "type": "string",
                    "example": "somebody@naver.com"
                },
                "keywords": {
                    "description": "회원 가입시 입력한 키워드",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "b0d88d67-01fd-47f8-b426-6ca0657d0f6e"
                    ]
                },
                "nickname": {
                    "description": "회원 가입시 입력한 닉네임",
                    "type": "string",
                    "example": "somebody"
                },
                "password": {
                    "description": "회원 가입시 입력한 패스워드",
                    "type": "string",
                    "example": "04F0BF9AB5963706415A0EC4C0E8167EBC8640B0B9A40870CDC39BE8A8DA813B"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}